requirements.txt
# requirements.txt
# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# YOLO
ultralytics>=8.0.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.0.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 机器学习工具
scikit-learn>=0.24.0

# 进度条
tqdm>=4.62.0

# 配置文件
PyYAML>=5.4.0